python/requirement.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asteroid-filterbanks==0.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
blinker==1.8.2
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
DateTime==5.5
decorator==5.1.1
diart==0.9.1
dnspython==2.6.1
docopt==0.6.2
einops==0.8.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpeg==1.4
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.5.4
google-api-core==2.19.1
google-auth==2.32.0
google-cloud==0.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.64.1
grpcio-tools==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
HyperPyYAML==1.2.2
idna==3.7
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning==2.3.3
lightning-utilities==0.11.5
llvmlite==0.42.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
namex==0.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
resampy==0.4.3
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
Rx==3.2.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.2
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websocket-server==0.6.4
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zope.interface==6.4.post2
